# javascript/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "0.1",
      "title": "Documentation de l'API",
      "description": "Swagger à pour but de documenter et tester l'API",
      "contact": {}
    },
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/forgot_password": {
        "post": {
          "description": "L'utilisateur peut déclarer la perte de son mot de passe. Il doit renseigner son pseudo et adresse e-mail pour recevoir un mot de passe provisoire par mail.",
          "summary": "Mot de passe oublié",
          "tags": [
            "auth"
          ],
          "operationId": "forgot_password",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "email",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": "Adresse mail"
            },
            {
              "name": "nickname",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": "Pseudo"
            }
          ],
          "responses": {
            "200": {
              "description": "Confirmation envoie de l'email",
              "schema": {
                "$ref": "#/definitions/responseApi"
              }
            },
            "500": {
              "description": "Erreur seveur",
              "schema": {
                "$ref": "#/definitions/errorApi"
              }
            },
            "400": {
                "description": "Retour d'erreur",
                "schema": {
                  "$ref": "#/definitions/errorApi"
                }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "errorApi": {
        "title": "Error API",
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "responseApi": {
        "title": "Response API",
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
...
}